split out TileInfoPanel from TileContextPanel
